==> 02_titanic_data_cleaning.ipynb <==
# Kaggle Competition: Titanic
Predicting whether an idividual survived or not

import numpy as np
import pandas as pd
pd.set_option('display.max_columns',100)

import matplotlib.pyplot as plt
%matplotlib inline

import seaborn as sns

_________________

# Titanic Data Cleaning 

df = pd.read_csv('titanic.csv')

df.shape

df.tail(3)

_________________

## Unwanted Observations

### Duplicates

Assuming that this dataset has kept all necessary records. There are duplicates but since the dataset from Kaggle also contains 891 records, I will not remove any duplicates. 

### Irrelevant

df.groupby('survived').alive.value_counts()

case1 = ((df.survived == 1) & (df.alive == 'yes')).astype(int).sum()
case2 = ((df.survived == 0) & (df.alive == 'no')).astype(int).sum()
print(case1)
print(case2)
print('both should add up to 891')
print('{} + {} = {}'.format(case1, case2, case1+case2))

The feature '*alive*' is the EXACT same as '*survived*', will drop 'alive'.

del df['alive']

df.shape

df.columns

'*Alive*' feature is gone.

df[['embarked','embark_town']].drop_duplicates()

'embarked' feature is the **EXACT** same as 'embark_town'. Remove 'embark_town' feature.

df.iloc[61,:]

del df['embark_town']

df.head(2)

_________________

## Structural Errors

### Indicator Variables

df.head()

indicator_features = ['survived', 'sex', 'adult_male', 'alone']

df.dtypes

df.survived.unique()

del indicator_features[0]
print(indicator_features)

df.rename(columns={'sex':'isFemale'},inplace=True)

df.isFemale.unique()

df.isFemale.replace('female',1,inplace=True)
print(df.head(3))

df.isFemale.replace('male',0,inplace=True)
print(df.head(3))

df.isFemale.unique()

del indicator_features[0]
print(indicator_features)

df.adult_male.unique()

df.adult_male.head()

df.adult_male = df.adult_male.astype(int)

df.adult_male.head()

del indicator_features[0]
print(indicator_features)

df.alone.unique()

df.alone.head()

df.alone = df.alone.astype(int)

df.alone.head()

del indicator_features[0]
print(indicator_features)

### Typos and Capitalizations

df.head()

categorical_features = [col for col in df.select_dtypes(include=[object]).columns]
print(categorical_features)

df.embarked.unique()

del categorical_features[0]
print(categorical_features)

df['class'].unique()

df['class'].tail()

#df['class'] = df['class'].replace(['Third','First', 'Second'],[3,1,2],inplace=True)
df['class'].replace(['Third','First', 'Second'],[3,1,2],inplace=True)

df['class'].tail()

del categorical_features[0]
print(categorical_features)

df.who.unique()

del categorical_features[0]
print(categorical_features)

df.deck.unique()

del categorical_features[0]
print(categorical_features)

### Mislabeled Classes

Not applicable in this case, the classes (categorical) features are few in number with few instances of each class. 

_________________

## Unwanted Outliers

Since this dataset comes from Kaggle, is a well used (beginner) dataset, and no huge anomalies were discovered during the Data Exploration phase of the project, I will be removing no outliers from the dataset.

_________________

## Missing Data

### Missing Categorical

categorical_features = [col for col in df.select_dtypes(include=[object]).columns]
print(categorical_features)

df[categorical_features].isnull().sum()

cat_missing = [col for col in categorical_features if df[col].isnull().sum() > 0]
print(cat_missing)

df.embarked.value_counts(normalize=True)

With the understanding that my change will make any machine learning algorithm not be able to attain 100% accuracy with any certainty, I will go ahead and manually updated the 2 missing 'embarked' features to 'S' since that class makes up 72.44% of the values.

df.embarked.fillna('S',inplace=True)

df[cat_missing].isnull().sum()

del cat_missing[0]
print(cat_missing)

df.deck.value_counts()

missing_deck = (df.deck.isnull().sum()/len(df.deck))
not_missing = (1-missing_deck)
print('{} non-null objects.\n'.format(df.deck.isnull().sum()))
print('That means {}% of values in deck are not null'.format(missing_deck*100))
print('{}% of the values are missing'.format(not_missing*100))

df.deck.fillna('Missing',inplace=True)

df[cat_missing].isnull().sum()

del cat_missing[0]
print(cat_missing)

### Missing Numeric

df.select_dtypes(include=[np.number]).isnull().sum()

df.select_dtypes(exclude=[object]).isnull().sum()

**Only 'age' feature has missing values.**

missing_age = (df.age.isnull().sum()/len(df.age))
not_missing = (1-missing_age)
print('{} non-null objects.\n'.format(df.age.isnull().sum()))
print('That means {}% of values in deck are not null'.format(missing_age*100))
print('{}% of the values are missing'.format(not_missing*100))

df[(df.age.isnull())][['age','who']].drop_duplicates()

m_age_m = df[df.who == 'man'].groupby('who').age.mean()[0].round()
m_age_w = df[df.who == 'woman'].groupby('who').age.mean()[0].round()
print(m_age_m)
print(m_age_w)

To me, there seems to be 3 options here: 
1. Drop the 'age' feature entirely. 
2. Drop the 177 observations that have a null value for the 'age' feature
3. Impute the average age for 'man' and 'woman' (only 'who' features that have missing values) and then replace the nulls with those average values.  
4. **Flag and fill the missing value. <-------- CHOSEN OPTION**

df['age_missing'] = df.age.isnull().astype(int)
print(df.tail(3))

df.age.fillna(0, inplace=True)
print(df.tail(3))

df.to_csv('titanic_cleaned.csv', index=None)
