==> 03_titanic_feature_engineering.ipynb <==
# Kaggle Competition: Titanic
Predicting whether an idividual survived or not

import numpy as np
import pandas as pd
pd.set_option('display.max_columns',100)

import matplotlib.pyplot as plt
%matplotlib inline

import seaborn as sns

_________________

# Titanic Feature Engineering

df = pd.read_csv('titanic_cleaned.csv')

df.shape

df.tail(3)

_________________

## Domain Knowledge 

df.tail(3)

### Boolean masks

Boolean masks allow you to easily create indicator variables based on conditionals. They empower
you to specifically isolate information that you suspect might be important for the model.

### Interaction Features

Will create a feature that is a calculation of the total party size for an individual. This will include the individual themself, meaning that the smallest group size is 1. 

df['group_size'] = 1 + df.sibsp + df.parch
df.tail(3)

##  Sparse Classes

df.select_dtypes(include=[object]).head(3)

### Similar Classes
There are few classes and they are unique enough so no need to group several classes together

### Other Classes

df.deck.value_counts()

Can group the 'A', 'F', 'G' values of the 'deck' feature together but for now, will leave them as is. 

## Dummy Variables

df.select_dtypes(include=[object]).head(3)

categorical_colums = [col for col in df.select_dtypes(include=[object]).columns]
print(categorical_colums)

df = pd.get_dummies(df, columns=categorical_colums)

df.head()

## Remove Unused/Redundant

There are no unused/redundant features here. 

# Save the analytic base table

df.isnull().sum()

df.to_csv('titanic_base_table.csv', index=None)
