==> 04_titanic_model_training.ipynb <==
# Kaggle Competition: Titanic
Predicting whether an idividual survived or not

import numpy as np
import pandas as pd
pd.set_option('display.max_columns',100)
import pickle

import matplotlib.pyplot as plt
%matplotlib inline

import seaborn as sns

from sklearn.model_selection import train_test_split

# Import Logistic Regression
from sklearn.linear_model import LogisticRegression
# Import RandomForestClassifier and GradientBoostingClassifer
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
# Function for splitting training and test set
from sklearn.model_selection import train_test_split
# Function for creating model pipelines
from sklearn.pipeline import make_pipeline
# For standardization
from sklearn.preprocessing import StandardScaler
# Helper for cross-validation
from sklearn.model_selection import GridSearchCV
# Classification metrics (added later)
from sklearn.metrics import roc_curve, auc, confusion_matrix

Steps to follow: 
1. [Split dataset into train and test](#data_split)
2. [Build the model pipelines](#pipeline)
3. [Declare hyperparameters to tune for our task](#hyperparameters)
4. [Fit and tune models using cross-validation](#tune)
5. [Evaluate model](#evaluate)
6. [Area under ROC curve](#auroc)

df = pd.read_csv('titanic_base_table.csv')

df.shape

df.tail(3)

_________________

<span id='data_split'> </span>
# Split Data into Test/Train Split

y = df.survived

X = df.drop('survived', axis=1)

print(y.shape)
print(X.shape)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.2, random_state=42,
                                                   stratify=df.survived)

print(len(X_train),len(X_test), len(y_train), len(y_test))

_________________

<span id='pipeline'></span>
# Build Pipelines

Models tested: 
* Logistic Regression w/ L1 Penalty
* Logistic Regression w/ L2 Penalty
* Random Forest
* Gradient Boosted Tree

pipelines = {
    'l1':make_pipeline(StandardScaler(), 
                       LogisticRegression(penalty='l1', random_state=42,n_jobs=-1)),
    'l2':make_pipeline(StandardScaler(), 
                       LogisticRegression(penalty='l2', random_state=42,n_jobs=-1)),
    'rf':make_pipeline(StandardScaler(),
                       RandomForestClassifier(random_state=42,n_jobs=-1)), 
    'gb':make_pipeline(StandardScaler(), GradientBoostingClassifier(random_state=42))
}

_________________

<span id='hyperparameters'></span>
# Declare Hyperparameters to Tune

pipelines['l1'].get_params()

l1_hyperparameters = {'logisticregression__C': np.linspace(1e-3, 1e3, 10)}
l2_hyperparameters = {'logisticregression__C': np.linspace(1e-3, 1e3, 10)}

pipelines['rf'].get_params()

rf_hyperparameters = {'randomforestclassifier__n_estimators': [50, 100, 200],
                     'randomforestclassifier__max_features': ['auto', 'sqrt', .33],
                     'randomforestclassifier__max_depth': [3, 5, 8]}

pipelines['gb'].get_params()

gb_hyperparameters = {'gradientboostingclassifier__n_estimators': [50, 100, 200],
                     'gradientboostingclassifier__learning_rate': [.05, .1, .2],
                     'gradientboostingclassifier__max_depth': [3,5,8],
                     'gradientboostingclassifier__max_features': ['auto','sqrt',.33]}

hyperparameters = {
    'l1': l1_hyperparameters,
    'l2': l2_hyperparameters,
    'rf': rf_hyperparameters, 
    'gb': gb_hyperparameters
}

_________________

<span id='tune'></span>
# Fit and Tune Models Using Cross Validation

%%time

fitted_models = {}

for name, pipeline in pipelines.items():
    #print('{}:: {}\n\n'.format(name, pipeline))
    
    # create cross validation object
    model = GridSearchCV(pipeline, hyperparameters[name], cv=10, n_jobs=-1)
    
    # fit the model
    model.fit(X_train, y_train)
    
    # store the model in the dictionary
    fitted_models[name] = model
    
    print('{} has been fitted.'.format(name))

_________________

<span id='evaluate'></span>
# Evaluate the Models

for k, v in fitted_models.items():
    print(k, v.best_score_)

fitted_models['rf']

_________________

<span id='auroc'></span>
# Area Under the ROC Curve

for name, model in fitted_models.items():
    pred = model.predict_proba(X_test)
    # get the 'survived' (1) class
    pred = [p[1] for p in pred]
    
    fpr, tpr, thresholds = roc_curve(y_test, pred)
    
    print('{} has an auc score of {}'.format(name, auc(fpr, tpr)))

## Logistic Regression w/ L1 penalty gave us the best AUC score
Save the L1 model as a pickle file

with open('titanic_final_model.pkl','wb') as f:
    pickle.dump(fitted_models['l1'].best_estimator_,f)    

_________________
